# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for a freelancer/service-provider business management app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-templates:
    get:
      operationId: listServiceTemplates
      tags: [service-templates]
      summary: List reusable service templates for the current team
      responses:
        "200":
          description: Service templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceTemplate"
    post:
      operationId: createServiceTemplate
      tags: [service-templates]
      summary: Create a reusable service template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTemplateInput"
      responses:
        "201":
          description: Service template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTemplate"
  /service-templates/{id}:
    get:
      operationId: getServiceTemplate
      tags: [service-templates]
      summary: Get a service template with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTemplate"
        "404":
          description: Service template not found
    patch:
      operationId: updateServiceTemplate
      tags: [service-templates]
      summary: Update a reusable service template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTemplateUpdate"
      responses:
        "200":
          description: Service template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTemplate"
        "404":
          description: Service template not found
    delete:
      operationId: deleteServiceTemplate
      tags: [service-templates]
      summary: Delete a reusable service template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Service template deleted
        "404":
          description: Service template not found
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for the current team
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (including marking it done)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
  /tasks/{id}/photos:
    post:
      operationId: addTaskPhoto
      tags: [tasks]
      summary: Attach an uploaded photo to a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPhotoInput"
      responses:
        "201":
          description: Photo attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskPhoto"
        "404":
          description: Task not found
  /tasks/{id}/photos/{photoId}:
    delete:
      operationId: deleteTaskPhoto
      tags: [tasks]
      summary: Remove a photo from a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Photo deleted
        "404":
          description: Task or photo not found
...
components:
  schemas:
...
    ServiceTemplateItem:
      type: object
      required: [id, productId, description, quantity, unitPrice]
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    ServiceTemplateItemInput:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        productId:
          type: ["integer", "null"]
        description:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unitPrice:
          type: number
          minimum: 0
    ServiceTemplate:
      type: object
      required:
        [id, name, serviceScopeEnabled, serviceDescription, notes, laborCost, items, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        serviceScopeEnabled:
          type: boolean
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        laborCost:
          type: number
          minimum: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTemplateItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServiceTemplateInput:
      type: object
      required: [name, items]
      properties:
        name:
          type: string
          minLength: 1
        serviceScopeEnabled:
          type: boolean
          default: false
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        laborCost:
          type: number
          minimum: 0
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceTemplateItemInput"
    ServiceTemplateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        serviceScopeEnabled:
          type: boolean
        serviceDescription:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        laborCost:
          type: number
          minimum: 0
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceTemplateItemInput"
...
    TaskStatus:
      type: string
      enum: [scheduled, in_progress, completed, paid]
    Task:
      type: object
      required:
        [
          id,
          title,
          description,
          dueAt,
          endAt,
          status,
          clientId,
          clientName,
          createdAt,
          paidAt,
          paidAmount,
          photos,
          quoteId,
        ]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueAt:
          type: string
          format: date-time
        endAt:
          type: ["string", "null"]
          format: date-time
          description: Data prevista de término do serviço.
        status:
          $ref: "#/components/schemas/TaskStatus"
        clientId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
          description: Quote that originated this task, when it was created from an approved quote.
        clientName:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        paidAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        photos:
          type: array
          items:
            $ref: "#/components/schemas/TaskPhoto"
    TaskInput:
      type: object
      required: [title, dueAt, endAt]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueAt:
          type: string
          format: date-time
        endAt:
          type: ["string", "null"]
          format: date-time
        clientId:
          type: ["integer", "null"]
...
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        dueAt:
          type: string
          format: date-time
        endAt:
          type: ["string", "null"]
          format: date-time
        status:
          $ref: "#/components/schemas/TaskStatus"
        clientId:
          type: ["integer", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        paidAmount:
          type: ["number", "null"]
    TaskPhoto:
      type: object
      required: [id, taskId, url, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    TaskPhotoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
...