# lib/api-zod/src/generated/types/serviceTemplateItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a freelancer/service-provider business management app
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceTemplateItem {
  id: number;
  /** @nullable */
  productId: number | null;
  description: string;
  quantity: number;
  unitPrice: number;
}