# lib/api-zod/src/generated/types/serviceTemplateUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a freelancer/service-provider business management app
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceTemplateItemInput } from './serviceTemplateItemInput';

export interface ServiceTemplateUpdate {
  /** @minLength 1 */
  name?: string;
  serviceScopeEnabled?: boolean;
  /** @nullable */
  serviceDescription?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @minimum 0 */
  laborCost?: number;
  /** @minItems 1 */
  items?: ServiceTemplateItemInput[];
}